backend/requirements.txt
Django==5.1
djangorestframework
django-cors-headers
psycopg2-binary
